partsomething/requirements.txt
astroid==2.5.6
bing-image-downloader==1.1.0
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
dlib==19.22.99
face-recognition==1.3.0
face-recognition-models==0.3.0
ffmpeg-python==0.2.0
Flask==2.0.2
future==0.18.2
idna==2.10
isort==5.8.0
itsdangerous==2.0.1
Jinja2==3.0.2
joblib==1.0.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
numpy==1.20.3
opencv-python==4.5.2.54
Pillow==8.2.0
pylint==2.8.3
requests==2.25.1
scikit-learn==0.24.2
scipy==1.6.3
svm==0.1.0
threadpoolctl==2.1.0
toml==0.10.2
urllib3==1.26.5
Werkzeug==2.0.2
wrapt==1.12.1
xmltodict==0.12.0
